feat(CP-639): using a single rule to check for granting basic roles (#16)
It replaces the two individual rules that focused on serviceAccount and Group principals specifically. This rule simply loops through all `google_` resources with a `role` attribute.

Restructured the `manual_tests` to make maintenance a bit easier. Since the initial `main.tf` file has grown over time it became difficult to determine if and when a rule should be triggered.

Also CP-304